chore: change folder structure
