--- src/App.tsx
@@ -1,13 +1,0 @@
-import "./App.css";
-import "../src/styles/global.css";
-import { HomePage } from "./pages/home/HomePage";
-
-function App() {
-  return (
-    <>
-      <HomePage />
-    </>
-  );
-}
-
-export default App;
