+++ src/features/home/components/events-section/__tests__/EventCard.test.tsx
