+++ src/features/home/components/events-section/__tests__/EventSection.test.tsx
