+++ src/features/search/SearchSection.tsx
@@ -1,9 +1,10 @@
 import { useState } from 'react';
 import styles from './SearchSection.module.css';
-import { Dropdown } from '../../components/ui/dropdown/Dropdown';
+
 import { FaSearch } from 'react-icons/fa';
 import { dates, eventTypes, locations } from './data/data';
-import { Button } from '../../components/ui/Button/Button';
+import { Dropdown } from '../../shared/ui/dropdown/Dropdown';
+import { Button } from '../../shared/ui/Button/Button';
 
 export const SearchSection = () => {
   const [searchParams, setSearchParams] = useState({
