+++ src/shared/hooks/useAuth.ts
@@ -1,0 +1,4 @@
+export const useAuth = () => {
+  const user = JSON.parse(localStorage.getItem('user') || 'null');
+  return { user, isAuthenticated: !!user };
+};
