+++ src/shared/hooks/useClickOutside.tsx
