+++ src/shared/ui/__tests__/Dropdown.test.tsx
@@ -1,6 +1,7 @@
 import { render, screen } from '@testing-library/react';
-import { Dropdown } from '../../../components/ui/dropdown/Dropdown';
+
 import userEvent from '@testing-library/user-event';
+import { Dropdown } from '../dropdown/Dropdown';
 
 describe('Dropdown', () => {
   const user = userEvent.setup();
